hyperboloid/hyperboloid_requirements.txt
jax==0.4.14
jaxlib==0.4.14
numpy==1.26.4
scipy==1.13.1
matplotlib==3.9.2
opencv-python==4.10.0.84
pyserial==3.5
aiohttp==3.10.5
flask==3.0.3
flask-socketio==5.3.6
torch==2.3.1+cu121
sentence-transformers==3.0.1
requests==2.31.0
python-dotenv==1.0.1
gitpython==3.1.43
schedule==1.2.0
